refactor board
