+++ src/modules/deals/components/DealItem.tsx
@@ -10,7 +10,7 @@ import {
   Status
 } from 'modules/boards/styles/item';
 import { Content } from 'modules/boards/styles/stage';
-import { renderAmount } from 'modules/boards/utils';
+import { renderAmount, renderPriority } from 'modules/boards/utils';
 import Icon from 'modules/common/components/Icon';
 import { __ } from 'modules/common/utils';
 import Participators from 'modules/inbox/components/conversationDetail/workarea/Participators';
@@ -129,7 +129,10 @@ class DealItem extends React.PureComponent<Props, { isPopupVisible: boolean }> {
       <>
         <Labels labels={item.labels} indicator={true} />
 
-        <h5>{item.name}</h5>
+        <h5>
+          {renderPriority(item.priority)}
+          {item.name}
+        </h5>
 
         <Details color="#63D2D6" items={products} />
         <Details color="#F7CE53" items={customers || []} />
